Authorization using gate

+++ app/Http/Controllers/QuestionsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\AskQuestionRequest;
 use App\Models\Question;
+use Illuminate\Auth\Access\Gate;
 use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
@@ -28,8 +29,8 @@ class QuestionsController extends Controller
      */
     public function create()
     {
-        $questions = new Question();
-        return view('questions.create', compact('questions'));
+        $question = new Question();
+        return view('questions.create', compact('question'));
     }
 
     /**
@@ -70,7 +71,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)){
+            abort(403, "Access denied"); //if not permitted
+        }
         return view('questions.edit', compact('question'));
+
     }
 
     /**
@@ -82,6 +87,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)){
+            abort(403, "Access denied"); //if not permitted
+        }
         $question->update($request->only('title', 'body'));
         session()->flash('success', 'Your question has been updated');
         return redirect()->route('questions.index');
@@ -95,6 +103,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)){
+            abort(403, "Access denied"); //if not permitted
+        }
         $question->delete();
         session()->flash('success', 'Your question has been deleted');
         return redirect()->route('questions.index');
